Move make_server to the root of the net unit tests crate.
