+++ tests/unit/net/lib.rs
@@ -35,6 +35,7 @@ extern crate util;
 
 use devtools_traits::DevtoolsControlMsg;
 use filemanager_thread::{TestProvider, TEST_PROVIDER};
+use hyper::server::{Handler, Listening, Server};
 use net::fetch::methods::{FetchContext, fetch};
 use net::filemanager_thread::FileManager;
 use net::test::HttpState;
@@ -44,6 +45,7 @@ use net_traits::response::Response;
 use std::rc::Rc;
 use std::sync::mpsc::Sender;
 use std::thread;
+use url::Url;
 
 const DEFAULT_USER_AGENT: &'static str = "Such Browser. Very Layout. Wow.";
 
@@ -78,4 +80,14 @@ fn fetch_async(request: Request, target: Box<FetchTaskTarget + Send>, dc: Option
 
 fn fetch_sync(request: Request, dc: Option<Sender<DevtoolsControlMsg>>) -> Response {
     fetch(Rc::new(request), &mut None, new_fetch_context(dc))
+}
+
+fn make_server<H: Handler + 'static>(handler: H) -> (Listening, Url) {
+    // this is a Listening server because of handle_threads()
+    let server = Server::http("0.0.0.0:0").unwrap().handle_threads(handler, 1).unwrap();
+    let port = server.socket.port().to_string();
+    let mut url_string = "http://localhost:".to_owned();
+    url_string.push_str(&port);
+    let url = Url::parse(&url_string).unwrap();
+    (server, url)
 }
